requirements.txt
anyio==4.4.0
APScheduler==3.6.3
asgiref==3.8.1
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==43.0.0
decorator==5.1.1
defusedxml==0.8.0rc2
Django==5.1
django-admin-list-filter-dropdown==1.0.3
django-cors-headers==4.3.1
django-filter==24.3
django-json-widget==2.0.1
django-templated-mail==1.1.1
django-unfold-admin==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-nested-routers==0.94.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jalali_core==1.0.0
jdatetime==5.0.0
oauthlib==3.2.2
pillow==10.4.0
psycopg2==2.9.9
pyaes==1.6.1
pyasn1==0.6.0
pycparser==2.22
PyJWT==2.9.0
pypng==0.20220715.0
python-telegram-bot==21.4
python3-openid==3.2.0
pytz==2024.1
qrcode==7.4.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
telegram-django-bot==1.1.0
Telethon==1.36.0
tornado==6.4
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
whitenoise==6.7.0
